+++ src/revocation.rs
@@ -362,7 +362,10 @@ impl CredentialStatus for RevocationList2020Status {
         };
         let revoked = match bitstring.get(credential_index) {
             Some(bitref) => *bitref,
-            None => false,
+            None => {
+                return result
+                    .with_error("Credential index in revocation list is invalid.".to_string());
+            }
         };
         if revoked {
             return result.with_error("Credential is revoked.".to_string());
@@ -478,7 +481,10 @@ impl CredentialStatus for RevocationList2021Status {
         };
         let revoked = match bitstring.get(credential_index) {
             Some(bitref) => *bitref,
-            None => false,
+            None => {
+                return result
+                    .with_error("Credential index in revocation list is invalid.".to_string());
+            }
         };
         if revoked {
             return result.with_error("Credential is revoked.".to_string());